Kbuild defconfig files.
This commit adds the defconfig files that define the standard
configurations for x86 and arm architecture and the standard
configurations for the following BSPs for microkernel and
nanokernel:

- x86 generic pc
- x86 quark
- arm fsl_frdm_k64f
- arm ti_lm3s6965

They can be applied using the make target defconfig.
This target will generate the corresponding .config
file that will be consumed by the build system.

Signed-off-by: Juan Manuel Cruz <juan.m.cruz.alcaraz@linux.intel.com>
Change-Id: I128ad52f792eeb61cc80578be5fc3bd427447aaa

Former-commit-id: 73c4417ac9567033aade3d273ad4a7d42a8f92db