Reduce floating point test duration
The length of time it took for the microkernel's test_fp_sharing test to run
was occasionally causing issues with sanity_chk.  The test would take close
to five minutes to finish and sanity_chk has a five minute timeout for the
tests.  To resolve this, the number of iterations for which the test executes
has been cut in half (from 1000 to 500).

Change-Id: I171b96f6af81a8b08d0367beecab536f16d1ca7e
Signed-off-by: Peter Mitsis <peter.mitsis@windriver.com>

Former-commit-id: e948817ee5fc54c5cb463ff7c5150ffd07319258