test_sema: test timeout APIs
Test the nano_xxx_sem_wait_timeout() APIs.

First, the task waits with a timeout and times out. Then it wait with a
timeout, but gets the semaphore in time.

Then, multiple timeout tests are done for the fibers, to test the
ordering of queueing/dequeueing when timeout occurs, first on one
semaphore, then on multiple semaphores.

Finally, multiple fibers pend on one semaphore, and they all get the
semaphore in time, except the last one: this tests that the timeout is
recomputed correctly when timeouts are aborted.

Change-Id: I73c174e6f15f38377a14573cb7f531e5e6ca4c07
Signed-off-by: Benjamin Walsh <benjamin.walsh@windriver.com>

Former-commit-id: 76103ecc2b858b705b0e74fc575c341251c49fd2