DOC: Restructures the documentation in two parts.
The contents are split in two: Zephyr Kernel and Zephyr Project. Under
Zephyr Kernel is all the content directly about the code base. Under
Zephyr Project is all the content about the development of the kernel
and of applications. Only the Doxygen document needed to be split. The
output of the in-code comments was placed under the Zephyr Kernel while
all the Doxygen guidelines and the extraction process information were
placed under the Zephyr Project.

Change-Id: I4f7b674d19449968e976b63f40bcaa9737cc0ecb
Signed-off-by: Rodrigo Caballero <rodrigo.caballero.abraham@intel.com>

Former-commit-id: 8fd2fd8de46e3f3a1403098b649b84584028a5a7